requirements.txt
# requirements.txt
# All Python packages needed for Project Samarth

requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.8
google-generativeai==0.3.0
faiss-cpu==1.7.4
streamlit==1.28.0
watchdog==2.3.1
